requirements.txt
<<<<<<< Updated upstream
flask
flask-restful
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-cors
flask-marshmallow
marshmallow-sqlalchemy
psycopg2-binary
python-dotenv
cloudinary
sendgrid
email-validator
pytest
pytest-cov
=======
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2026.4.22
cffi==1.17.1
cloudinary==1.44.2
coverage==7.6.1
cryptography==46.0.7
dj-database-url==3.0.1
Django==4.2.30
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
email-validator==2.3.0
exceptiongroup==1.3.1
gunicorn==23.0.0
idna==3.13
iniconfig==2.1.0
MarkupSafe==2.1.5
packaging==26.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
sendgrid==6.12.5
sib-api-v3-sdk==7.6.0
six==1.17.0
sqlparse==0.5.5
tomli==2.4.1
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==3.0.6
whitenoise==6.7.0
>>>>>>> Stashed changes
